pyproject.toml
[tool.poetry]
name = "telegram_bot"
version = "0.1.0"
description = ""
authors = ["Marco Silva <marcosilva0000+telbot@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
python-telegram-bot = "^11.1"
python-dotenv = "^0.10.3"
Google-Images-Search = "^1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
